fix(frameCache): change loop index type from byte to int

+++ Desktop/Services/FrameCache.cs
@@ -21,7 +21,7 @@ namespace Desktop
             if (_frames != null) return _frames;
 
             _frames = new List<BitmapImage>();
-            for (byte i = 1; i <= _frameQuantity; i++)
+            for (int i = 1; i <= _frameQuantity; i++)
             {
                 _frames.Add(CreateFrame(i));
             }
@@ -35,7 +35,7 @@ namespace Desktop
             await Task.Run(() =>
             {
                 var frames = new List<BitmapImage>();
-                for (byte i = 1; i <= _frameQuantity; i++)
+                for (int i = 1; i <= _frameQuantity; i++)
                 {
                     frames.Add(CreateFrame(i));
                 }
@@ -43,7 +43,7 @@ namespace Desktop
             });
         }
 
-        private BitmapImage CreateFrame(byte index)
+        private BitmapImage CreateFrame(int index)
         {
             var bitmap = new BitmapImage();
             bitmap.BeginInit();
